Split receipt-consuming-transaction to utils and tests
